updated enemy class with working takemdage function

+++ src/enemy.h
@@ -7,7 +7,7 @@ int hp, atk, def;
 bool isHostile; //for merchants and dragons
 public:
 	Enemy(char);
-	int takeDamage(int);
+	double takeDamage(int);
 	int getHp();
 	int getAtk();
 	int getDef();
